+++ src/main/java/fragment/ICodeFragment.java
@@ -2,7 +2,7 @@ package fragment;
 
 public interface ICodeFragment {
 
-	int getFragmentNumber();
+	long getFragmentNumber();
 
 	String getPath();
 }
